Default to building with swift.
Motivated by the absence of https://github.com/microsoft/vscode-cmake-tools/issues/3257
and to minimize developer confusion.
The contents of the tasks.json file in this commit are what VSCode
auto-populated, minus a `cwd` value that included my username.

Also drop the .gitignore'ing of the .vscode directory, since this is
something we do want developers to share.